pyproject.toml
[project]
name = "fastapi-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.0",
    "asyncpg>=0.31.0",
    "email-validator>=2.3.0",
    "fastapi>=0.128.0",
    "fastapi-offline>=1.7.5",
    "opentelemetry-api>=1.39.1",
    "opentelemetry-instrumentation-fastapi>=0.60b1",
    "opentelemetry-instrumentation-sqlalchemy>=0.60b1",
    "opentelemetry-sdk>=1.39.1",
    "prometheus-client>=0.24.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "sqlalchemy>=2.0.45",
    "tenacity>=9.1.2",
    "uvicorn[standard]>=0.40.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "isort>=7.0.0",
    "pytest>=9.0.2",
    "ruff>=0.14.11",
]
